refactor(cli,mcp): regroup query-insights into queries and sql-pools
Dissolve the `query-insights` CLI group in favour of user-facing groupings:
- `query-insights request-history` → `queries request-history`
- `query-insights session-history` → `queries session-history`
- `query-insights frequent`        → `queries frequent`
- `query-insights long-running`    → `queries long-running`
- `query-insights pool-insights`   → `sql-pools insights`

MCP tool names are unchanged (`list_request_history`, `list_session_history`,
`list_frequent_queries`, `list_long_running_queries`, `list_sql_pool_insights`)
and the total tool count stays 73.  The shared service layer
(`services/query_insights.py`) is untouched.
