+++ src/fabric_dw/cli/_main.py
@@ -14,7 +14,6 @@ from fabric_dw.cli.commands.completion import completion_group
 from fabric_dw.cli.commands.config import config_group
 from fabric_dw.cli.commands.procedures import procedures_group
 from fabric_dw.cli.commands.queries import queries_group
-from fabric_dw.cli.commands.query_insights import query_insights_group
 from fabric_dw.cli.commands.restore_points import restore_points_group
 from fabric_dw.cli.commands.schemas import schemas_group
 from fabric_dw.cli.commands.snapshots import snapshots_group
@@ -87,7 +86,6 @@ cli.add_command(warehouses_group)
 cli.add_command(sql_endpoints_group)
 cli.add_command(audit_group)
 cli.add_command(queries_group)
-cli.add_command(query_insights_group)
 cli.add_command(restore_points_group)
 cli.add_command(snapshots_group)
 cli.add_command(sql_group)
