+++ src/fabric_dw/mcp/tools/sql_pools.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import logging
+from datetime import datetime
 from typing import Annotated, Any
 
 from mcp.server.fastmcp import FastMCP
@@ -16,8 +17,9 @@ from fabric_dw.mcp._guards import (
     assert_workspace_allowed,
     assert_writes_allowed,
 )
-from fabric_dw.mcp._helpers import fabric_err
+from fabric_dw.mcp._helpers import fabric_err, make_sql_target, resolve_item
 from fabric_dw.models import SqlPool, SqlPoolClassifier
+from fabric_dw.services import query_insights as _qi_svc
 from fabric_dw.services import sql_pools as sql_pools_svc
 
 __all__ = ["register"]
@@ -289,3 +291,45 @@ def register(mcp: FastMCP) -> None:  # noqa: PLR0915
         except FabricError as exc:
             raise fabric_err(exc) from exc
         return result.model_dump(by_alias=True, mode="json")
+
+    def _parse_dt(value: str | None, param: str) -> datetime | None:
+        """Parse an ISO-8601 string to datetime, raising ToolError on bad input."""
+        if value is None:
+            return None
+        try:
+            return datetime.fromisoformat(value)
+        except ValueError as exc:
+            raise ToolError(f"invalid {param} {value!r}: expected ISO-8601") from exc
+
+    @mcp.tool(name="list_sql_pool_insights")
+    async def list_sql_pool_insights(
+        workspace: str,
+        warehouse: str,
+        limit: Annotated[int, Field(ge=1, le=10000)] = 100,
+        since: str | None = None,
+        until: str | None = None,
+    ) -> list[dict[str, Any]]:
+        """Return SQL pool insight events from queryinsights.sql_pool_insights.
+
+        Args:
+            workspace: Workspace name or GUID.
+            warehouse: Warehouse or SQL Analytics Endpoint name or GUID.
+            limit: Maximum rows to return (1-10000, default 100).
+            since: Optional ISO-8601 lower bound on timestamp.
+            until: Optional ISO-8601 upper bound on timestamp.
+        """
+        since_dt = _parse_dt(since, "since")
+        until_dt = _parse_dt(until, "until")
+        assert_workspace_allowed(workspace)
+        ctx = get_context()
+        try:
+            ws_id, item = await resolve_item(ctx.resolver, workspace, warehouse)
+            assert_workspace_allowed(workspace, str(ws_id))
+            _log.debug("list_sql_pool_insights ws=%s item=%s limit=%d", ws_id, item.id, limit)
+            target = make_sql_target(ws_id, item, warehouse)
+            result = await _qi_svc.list_sql_pool_insights(
+                target, limit=limit, since=since_dt, until=until_dt, mode=ctx.auth_mode
+            )
+        except FabricError as exc:
+            raise fabric_err(exc) from exc
+        return [q.model_dump(by_alias=True, mode="json") for q in result]
